pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.57.0",
    "gradio>=6.3.0",
    "httpx>=0.28.1",
    "openai>=2.15.0",
    "openai-agents>=0.6.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
